backend/zoo_project/pyproject.toml
[tool.poetry]
name = "zoo-project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
django-environ = "^0.11.2"
django-extensions = "^3.2.3"
djangorestframework = "^3.15.1"
whitenoise = "^6.6.0"
psycopg = "^3.1.19"
psycopg-binary = "^3.1.19"
django-debug-toolbar = "^5.0.1"
django-cors-headers = "^4.7.0"

[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
ipython = "^8.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
